Fix #128 - Cannot attach raw block volumes
CSI has supported raw block volumes since version 0.1, and Kubernetes
since v1.9, but the external-attacher cannot attach a block volume.

When attaching a raw block volume we don't even see an error on the
external-attacher logs saying it cannot handle it, it just goes ahead
assuming it's a mount type of volume.

This patch checks the volume mode on the spec to see if it's a raw
volume or not.

The the volume mode is define and it's PVBlock we'll use the Block
access type, in any other case (undefined and Filesystem mode) we'll use
the Mount access type.