replace "dep" with "go mod"
This is the result of:
    go mod init
    release-tools/go-get-kubernetes.sh 1.15.4
    go get k8s.io/csi-translation-lib@17db17aaadeb70b1d7259c18ff045f51a1164d39
    # manually remove 'replace' statement for k8s.io/csi-translation-lib in go.mod
    go mod tidy
    go mod vendor

In other words, k8s.io/csi-translation-lib is kept at the same
revision as before because that's what the code is compatible with
whereas the rest of the Kubernetes modules get updated to 1.15.4.

A consistent update to 1.15.4 or even 1.16.0 can follow later, but
then will have to include source code changes.