allow setting default fs type through command line argument
The external attacher was using a hardcoded default ext4 filesystem
which was applied when PV spec did not contain fstype.

As a result attacher always reported ext4 as fs_type volume
capability creating interference with some CSI drivers.

This patch adds a flag to csi-attacher binary which provides users with
a workaround to set their own default filesystem or simply disable the
interference by using an empty string:

./bin/csi-attacher --default-fstype ""

When the flag is not used  we still use ext4, next major release will
allow behaviour change and we can default to "" instead preserving the
argument in case it's ever needed.