Rework order of APi calls after successful detach
Remove a finalizer on a VolumeAttachment (VA) first and then mark its state as
detached (if the object still exists).

This ensures that ControllerUnpublish is called only once. It won't be
called on VA without the finalizer, even if it has `attached: true`.

The other way around, i.e. marking the VA as detached first and removing
the finalizer later, it can happen that the attacher calls
ControllerUnpublish twice. A VA with the finalizer, deletionTimestamp and
`attacher: false` is considered as "uncertainly attached" - a previous
ControllerPublish might have timed out and the attacher needs to confirm
the volume is detached with ControllerUnpublish.